Main menu, options, and control menu created
added a scene for the main menu. i can add and remove buttons with no problems.

the options scene has a button for settings, this button doesn't work yet. this is where audio controls and brightness can be modified for the player.

the options scene also has a button for the controls. i'll list all the game controls here for players to find over time.

next i'll make a pause menu which connects to the main menu and options scenes.